feat(fs, core): modernize drivers, optimize ZIP trailing cleanup, and consolidate test architecture
- Harden NTFS, EXT, FAT, and ISO drivers and align specifications

- Optimize ZIP trailing storage cleanup with bounded memory and flush tracking

- Consolidate test architecture by inlining unit tests into source files

- Update docs and remove obsolete publication scripts

+++ rimfs-core/src/meta.rs
@@ -1,5 +1,7 @@
 // SPDX-License-Identifier: MIT
 
+//! Core filesystem metadata contracts and geometry definitions.
+
 #[cfg(all(not(feature = "std"), feature = "alloc"))]
 use ::alloc::string::String;
 
@@ -9,7 +11,7 @@ pub use crate::utils::volume::*;
 /// Provides access to static metadata needed during formatting, allocation, injection, or checking.
 pub trait FsMeta<Unit: Ord + Copy> {
     /// Size of one allocation unit in bytes.
-    fn unit_size(&self) -> usize;
+    fn unit_size(&self) -> u64;
 
     /// Compute the offset (in bytes) on disk corresponding to a given allocation unit.
     fn unit_offset(&self, unit: Unit) -> u64;
@@ -24,7 +26,7 @@ pub trait FsMeta<Unit: Ord + Copy> {
     fn last_data_unit(&self) -> Unit;
 
     /// Total number of allocatable units.
-    fn total_units(&self) -> usize;
+    fn total_units(&self) -> u64;
 
     /// Total size in bytes of the FS.
     fn size_bytes(&self) -> u64;
